getTotalStockRemaining: accumulate in long
An int accumulator wrapped negative past 2^31 units, while qtyInStock
is read as long. Accumulating in long keeps the sum exact; the return
stays double.

+++ core/src/main/java/com/squarespace/template/plugins/platform/CommerceUtils.java
@@ -311,7 +311,9 @@ public class CommerceUtils {
     if (EnumSet.of(ProductType.DIGITAL, ProductType.GIFT_CARD).contains(type)) {
       return Double.POSITIVE_INFINITY;
     } else {
-      int total = 0;
+      // Accumulate in long: qtyInStock is read as long, and an int
+      // accumulator would wrap past 2^31 units. Return widens to double.
+      long total = 0;
       JsonNode variants = structuredContent.path("variants");
       for (int i = 0; i < variants.size(); i++) {
         JsonNode variant = variants.get(i);
